gestion des documents suite et fin ...

+++ src/main/java/com/immobilier/gestionImmobiliere/modules/statistiques/dto/GainAgenceDTO.java
@@ -3,9 +3,11 @@ package com.immobilier.gestionImmobiliere.modules.statistiques.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data @Builder
 public class GainAgenceDTO {
     private String periodeMois;   // ex: "Octobre 2026"
-    private Double totalCommissions;
-    private Double totalReverseAuxBailleurs;
+    private BigDecimal totalCommissions;
+    private BigDecimal totalReverseAuxBailleurs;
 }
